WIP: T01 -- Split telemetry_models.py into domain files
