+++ src/hassette/core/telemetry/registration_queries.py
@@ -3,8 +3,9 @@
 from typing import TYPE_CHECKING, Any
 
 from hassette.core.telemetry.helpers import row_to_dict, since_clause, source_tier_clause
+from hassette.schemas.job_models import JobSummary
+from hassette.schemas.listener_models import ListenerSummary, SlowHandlerRecord
 from hassette.schemas.query_constants import DEFAULT_QUERY_LIMIT
-from hassette.schemas.telemetry_models import JobSummary, ListenerSummary, SlowHandlerRecord
 from hassette.types.types import QuerySourceTier
 
 if TYPE_CHECKING:
