+++ src/hassette/schemas/__init__.py
@@ -22,54 +22,20 @@ from hassette.schemas.domain_models import (
 )
 from hassette.schemas.live_counts import LiveCounts
 from hassette.schemas.query_constants import DEFAULT_QUERY_LIMIT, DEFAULT_SPARKLINE_BUCKETS
-from hassette.schemas.telemetry_models import (
-    ActivityFeedEntry,
-    AppHealthSummary,
-    AppLastError,
-    BlockingEvent,
-    Execution,
-    GlobalSummary,
-    HandlerErrorRecord,
-    JobErrorRecord,
-    JobGlobalStats,
-    JobSummary,
-    ListenerGlobalStats,
-    ListenerSummary,
-    LogRecord,
-    SessionRecord,
-    SessionSummary,
-    SlowHandlerRecord,
-)
 
 __all__ = [
     "DEFAULT_QUERY_LIMIT",
     "DEFAULT_SPARKLINE_BUCKETS",
-    "ActivityFeedEntry",
     "AppFullSnapshot",
-    "AppHealthSummary",
     "AppInstanceInfo",
-    "AppLastError",
     "AppManifestInfo",
     "AppStatusChangedData",
     "AppStatusSnapshot",
-    "BlockingEvent",
     "BootIssue",
     "ConnectivityData",
-    "Execution",
-    "GlobalSummary",
-    "HandlerErrorRecord",
-    "JobErrorRecord",
-    "JobGlobalStats",
-    "JobSummary",
-    "ListenerGlobalStats",
-    "ListenerSummary",
     "LiveCounts",
-    "LogRecord",
     "ServiceInfo",
     "ServiceStatusData",
-    "SessionRecord",
-    "SessionSummary",
-    "SlowHandlerRecord",
     "StateChangedData",
     "SystemStatus",
 ]
