+++ src/hassette/web/utils.py
@@ -3,7 +3,7 @@
 from logging import getLogger
 from typing import TYPE_CHECKING
 
-from hassette.schemas.telemetry_models import JobSummary
+from hassette.schemas.job_models import JobSummary
 
 if TYPE_CHECKING:
     from hassette.core.scheduler_service import SchedulerService
