+++ tests/integration/telemetry/test_health_aggregates_and_global_listeners.py
@@ -13,7 +13,7 @@ import pytest
 
 from hassette.core.database_service import DatabaseService
 from hassette.core.telemetry.query_service import AppHealthAggregates, TelemetryQueryService
-from hassette.schemas.telemetry_models import ListenerSummary
+from hassette.schemas.listener_models import ListenerSummary
 
 from .helpers import (
     BASE_TS,
