+++ tests/integration/telemetry/test_telemetry_query_service_aggregates.py
@@ -12,9 +12,7 @@ import pytest
 from hassette.const.misc import SECONDS_PER_DAY
 from hassette.core.database_service import DatabaseService
 from hassette.core.telemetry.query_service import TelemetryQueryService
-from hassette.schemas.telemetry_models import (
-    AppHealthSummary,
-)
+from hassette.schemas.summary_models import AppHealthSummary
 
 from .helpers import (
     BASE_TS,
