+++ tests/integration/telemetry/test_telemetry_query_service_misc.py
@@ -13,9 +13,7 @@ from hassette.core.database_service import DatabaseService
 from hassette.core.telemetry.helpers import source_tier_clause
 from hassette.core.telemetry.query_service import TelemetryQueryService
 from hassette.exceptions import TelemetryUnavailableError
-from hassette.schemas.telemetry_models import (
-    ActivityFeedEntry,
-)
+from hassette.schemas.execution_models import ActivityFeedEntry
 from hassette.test_utils.mock_hassette import make_mock_hassette
 
 from .helpers import (
