+++ tests/unit/core/test_log_records.py
@@ -15,7 +15,7 @@ from hassette.config.models import LoggingConfig
 from hassette.core.database_service import DatabaseService
 from hassette.core.migration_runner import run_migrations
 from hassette.core.telemetry.query_service import TelemetryQueryService
-from hassette.schemas.telemetry_models import LogRecord
+from hassette.schemas.log_models import LogRecord
 
 from .conftest import TELEMETRY_TEST_DDL as DDL
